Safari-609.1.11.1/LayoutTests/http/tests/websocket/tests/hybi/error-event-ready-state-non-existent-url-with-server-responding-404-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8000/websocket/tests/hybi/resources/status-404-without-body.php' failed: Unexpected response code: 404
Test that readyState is CLOSED within onerror event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onerror() was called.
PASS ws.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
